Tests fixes after CurrentUser refactoring (#3042)
* Tests fixes after CurrentUser refactoring

* Added user_settings_store test
